+++ src/core/general.js
