+++ src/core/index.js
@@ -2,9 +2,9 @@
 const init = ({ lib }) => {
     const core = {
         constants: require('./constants'),
-        generalUtils: require('./general-utils'),
-        positionUtils: require('./position-utils').init({ lib }),
-        textUtils: require('./text-utils').init({ lib }),
+        general: require('./general'),
+        position: require('./position').init({ lib }),
+        text: require('./text').init({ lib }),
     }
 
     return core;
