+++ src/core/text.js
