+++ src/utils/geometry.js
